requirements.txt
kafka-python==2.0.2
apache-flink==1.15.0
google-cloud==2.10.0
apache-airflow==2.2.2
pandas==1.3.3
scikit-learn==0.24.2
matplotlib==3.4.3
pyarrow==5.0.0
tableau-api-lib==0.5.0
